+++ mpglib/layer3.c
@@ -737,7 +737,7 @@ III_dequantize_sample(PMPSTR mp, real xr[SBLIMIT][SSLIMIT], int *scf,
     {
         int     i;
 
-        for (i = (&xr[SBLIMIT][0] - xrpnt) >> 1; i > 0; i--) {
+        for (i = (int) (&xr[SBLIMIT][0] - xrpnt) >> 1; i > 0; i--) {
             *xrpnt++ = 0.0;
             *xrpnt++ = 0.0;
         }
@@ -1808,13 +1808,13 @@ decode_layer3_frame(PMPSTR mp, unsigned char *pcm_sample, int *pcm_point,
             }
             if (mp->pinfo != NULL) {
                 int     i;
-                mp->pinfo->sfbits[gr][0] = part2bits;
+                mp->pinfo->sfbits[gr][0] = (int)part2bits;
                 for (i = 0; i < 39; i++)
                     mp->pinfo->sfb_s[gr][0][i] = scalefacs[0][i];
             }
 
             /* lame_report_fnc(mp->report_err, "calling III dequantize sample 1 gr_infos->part2_3_length %d\n", gr_infos->part2_3_length); */
-            if (III_dequantize_sample(mp, hybridIn[0], scalefacs[0], gr_infos, sfreq, part2bits))
+            if (III_dequantize_sample(mp, hybridIn[0], scalefacs[0], gr_infos, sfreq, (int)part2bits))
                 return clip;
         }
         if (stereo == 2) {
@@ -1827,13 +1827,13 @@ decode_layer3_frame(PMPSTR mp, unsigned char *pcm_sample, int *pcm_point,
             }
             if (mp->pinfo != NULL) {
                 int     i;
-                mp->pinfo->sfbits[gr][1] = part2bits;
+                mp->pinfo->sfbits[gr][1] = (int) part2bits;
                 for (i = 0; i < 39; i++)
                     mp->pinfo->sfb_s[gr][1][i] = scalefacs[1][i];
             }
 
             /* lame_report_fnc(mp->report_err, "calling III dequantize sample 2  gr_infos->part2_3_length %d\n", gr_infos->part2_3_length); */
-            if (III_dequantize_sample(mp, hybridIn[1], scalefacs[1], gr_infos, sfreq, part2bits))
+            if (III_dequantize_sample(mp, hybridIn[1], scalefacs[1], gr_infos, sfreq, (int)part2bits))
                 return clip;
 
             if (ms_stereo) {
